pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "personal-ai-assistant"
version = "0.1.0"
description = "A personal AI operating layer for orchestrating tasks across life domains"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "psycopg[binary]>=3.2.0",
    "pyautogui>=0.9.54",
    "openai-whisper>=20231117",
    "jsonschema>=4.21.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pynput>=1.7.6",
    "openwakeword>=0.6.0",
    "langsmith>=0.3.0",
]

[project.optional-dependencies]
vector = [
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
